Add error handler for mutation classification
